+++ src/core/metadata.ts
@@ -1,4 +1,9 @@
-import {getGlobal, isFunction} from './utils';
+import {
+    getGlobal,
+    isFunction,
+    isJsObject,
+    isString
+} from './utils';
 import {PotionBase} from './potion';
 import {Item} from './item';
 
@@ -17,21 +22,44 @@ const Reflect = global.Reflect; // tslint:disable-line: variable-name
 })();
 
 
-const POTION_METADATA_KEY = Symbol('potion');
-export function potionInstance(ctor: typeof Item): PotionBase {
-    return Reflect.getOwnMetadata(POTION_METADATA_KEY, ctor);
+const potionInstanceMetadataKey = Symbol('potion:instance');
+const potionURIMetadataKey = Symbol('potion:uri');
+const potionPromiseCtorMetadataKey = Symbol('potion:promise');
+const potionReadonlyMetadataKey = Symbol('potion:readonly');
+const potionAsyncMetadataKey = Symbol('potion:async');
+
+
+/**
+ * Get the Potion instance from an Item constructor
+ * @param ctor
+ */
+export function getPotionInstance(ctor: typeof Item): PotionBase {
+    return Reflect.getOwnMetadata(potionInstanceMetadataKey, ctor);
 }
-export function decorateCtorWithPotionInstance(ctor: typeof Item, instance: any): void {
-    Reflect.defineMetadata(POTION_METADATA_KEY, instance, ctor);
+/**
+ * Set the Potion instance on an Item constructor
+ * @param ctor
+ * @param instance
+ */
+export function setPotionInstance(ctor: typeof Item, instance: PotionBase): void {
+    Reflect.defineMetadata(potionInstanceMetadataKey, instance, ctor);
 }
 
 
-const POTION_URI_METADATA_KEY = Symbol('potion:uri');
-export function potionURI(ctor: typeof Item): string {
-    return Reflect.getOwnMetadata(POTION_URI_METADATA_KEY, ctor);
+/**
+ * Get the Item uri from the Item constructor
+ * @param ctor
+ */
+export function getPotionURI(ctor: typeof Item): string {
+    return Reflect.getOwnMetadata(potionURIMetadataKey, ctor);
 }
-export function decorateCtorWithPotionURI(ctor: typeof Item, uri: string): void {
-    Reflect.defineMetadata(POTION_URI_METADATA_KEY, uri, ctor);
+/**
+ * Set the Item uri
+ * @param ctor
+ * @param uri
+ */
+export function setPotionURI(ctor: typeof Item, uri: string): void {
+    Reflect.defineMetadata(potionURIMetadataKey, uri, ctor);
 }
 
 
@@ -39,55 +67,125 @@ export function decorateCtorWithPotionURI(ctor: typeof Item, uri: string): void 
  * Get/Set the Promise implementation that should be used by Potion.
  * NOTE: If it is never set, it will fallback to using the native implementation of Promise.
  * @example
+ * @PotionConfig({
+ *     promiseCtor: Promise
+ * })
  * class Potion extends PotionBase {
  *     ...
  * }
- * setPotionPromise(Potion, ... some impl. of a Promise);
  */
-const POTION_PROMISE_METADATA_KEY = Symbol('potion:promise');
-export function potionPromise(potion: PotionBase): typeof Promise {
-    return Reflect.getOwnMetadata(POTION_PROMISE_METADATA_KEY, potion.constructor) || Promise;
+
+export function getPotionPromiseCtor(potion: PotionBase): typeof Promise {
+    return Reflect.getOwnMetadata(potionPromiseCtorMetadataKey, potion.constructor) || Promise;
 }
-export function setPotionPromise(ctor: typeof PotionBase, promise: any): void {
-    Reflect.defineMetadata(POTION_PROMISE_METADATA_KEY, promise, ctor);
+
+export function setPotionPromiseCtor(target: typeof PotionBase, promiseCtor: any): void {
+    Reflect.defineMetadata(potionPromiseCtorMetadataKey, promiseCtor, target);
 }
 
 
-const READONLY_METADATA_KEY = Symbol('potion:readonly');
-export function isReadonly(ctor: any, key: string): boolean {
-    const metadata = Reflect.getOwnMetadata(READONLY_METADATA_KEY, ctor);
-    return metadata && metadata[key];
+/**
+ * Check if a resource property is readonly
+ */
+export function isReadonly<T extends Item>(item: T, key: keyof T): boolean {
+    const ctors = getCtor(item.constructor as typeof Item);
+    const metadata = ctors.map(ctor => Reflect.getOwnMetadata(potionReadonlyMetadataKey, ctor));
+
+    for (const meta of metadata) {
+        if (isJsObject(meta)) {
+            const isReadonly = key && meta[key];
+            if (isReadonly) {
+                return true;
+            }
+        }
+    }
+
+    return false;
 }
 
 /**
  * Mark a resource property as readonly and omit when saved.
- *
  * @example
  * class User extends Item {
  *     @readonly
  *     age;
  * }
  */
-export function readonly(target: any, property: string): void {
+export function readonly(target: object, property: string): void {
+    decorate(target, potionReadonlyMetadataKey, property);
+}
+
+
+/**
+ * Check if a resource property is async
+ */
+export function isAsync(cls: typeof Item, keyOrUri: string): boolean {
+    const ctors = getCtor(cls);
+    const metadata = ctors.map(ctor => Reflect.getOwnMetadata(potionAsyncMetadataKey, ctor));
+
+    // Look up the prototype chain for this property and check if it is set
+    if (isString(keyOrUri)) {
+        for (const meta of metadata) {
+            if (isJsObject(meta)) {
+                const key = Object.keys(meta)
+                    .find(key => key === keyOrUri || keyOrUri.includes(key));
+                const isAsync = key && meta[key];
+                if (isAsync) {
+                    return true;
+                }
+            }
+        }
+    }
+
+    return false;
+}
+
+
+/**
+ * Walk the prototype chain and register all the constructors
+ * @param cls
+ */
+function getCtor(cls: typeof Item): Array<typeof Item> {
+    const ctors = [cls];
+    let ctor = Object.getPrototypeOf(cls);
+    while (ctor.prototype) {
+        ctors.push(ctor);
+        ctor = Object.getPrototypeOf(ctor);
+    }
+    return ctors;
+}
+
+
+/**
+ * Mark a resource property as async.
+ * @example
+ * class User extends Item {
+ *     @async
+ *     age: Promise<Foo>;
+ * }
+ */
+export function async(target: object, property: string): void {
+    decorate(target, potionAsyncMetadataKey, property);
+}
+
+
+/**
+ * Helper fn for decorating class properties
+ * @param target
+ * @param key
+ * @param property
+ */
+function decorate(target: object, key: symbol, property: string): void {
     const constructor = isFunction(target)
         ? target
         : isFunction(target.constructor)
             ? target.constructor
             : null;
-
-    if (constructor === null) {
-        // TODO: maybe throw an error here
-        return;
+    if (constructor) {
+        const metadata = Reflect.getOwnMetadata(key, constructor);
+        Reflect.defineMetadata(key, {
+            ...metadata,
+            [property]: true
+        }, constructor);
     }
-
-    Reflect.defineMetadata(
-        READONLY_METADATA_KEY,
-        Object.assign(
-            Reflect.getOwnMetadata(READONLY_METADATA_KEY, constructor) || {},
-            {
-                [property]: true
-            }
-        ),
-        constructor
-    );
 }